remove question renderers
