--- server/app/views/questiontypes/IdQuestionRenderer.java
@@ -1,54 +1,0 @@
-package views.questiontypes;
-
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import j2html.tags.specialized.DivTag;
-import services.Path;
-import services.applicant.ValidationErrorMessage;
-import services.applicant.question.ApplicantQuestion;
-import services.applicant.question.IdQuestion;
-import views.components.FieldWithLabel;
-
-/** Renders an id question. */
-public class IdQuestionRenderer extends ApplicantSingleQuestionRenderer {
-
-  public IdQuestionRenderer(ApplicantQuestion question) {
-    super(question);
-  }
-
-  @Override
-  public String getReferenceClass() {
-    return "cf-question-id";
-  }
-
-  @Override
-  protected DivTag renderInputTag(
-      ApplicantQuestionRendererParams params,
-      ImmutableMap<Path, ImmutableSet<ValidationErrorMessage>> validationErrors,
-      ImmutableList<String> ariaDescribedByIds,
-      boolean isOptional) {
-    IdQuestion idQuestion = applicantQuestion.createIdQuestion();
-
-    FieldWithLabel idField =
-        FieldWithLabel.input()
-            .setFieldName(idQuestion.getIdPath().toString())
-            .setValue(idQuestion.getIdValue().orElse(""))
-            .setAriaRequired(!isOptional)
-            .setFieldErrors(
-                params.messages(),
-                validationErrors.getOrDefault(idQuestion.getIdPath(), ImmutableSet.of()))
-            .setAriaDescribedByIds(ariaDescribedByIds)
-            .setScreenReaderText(applicantQuestion.getQuestionTextForScreenReader());
-
-    if (params.autofocusSingleField()) {
-      idField.focusOnInput();
-    }
-
-    if (!validationErrors.isEmpty()) {
-      idField.forceAriaInvalid();
-    }
-
-    return idField.getInputTag();
-  }
-}
