--- server/test/views/questiontypes/ApplicantQuestionRendererFactoryTest.java
@@ -1,80 +1,0 @@
-package views.questiontypes;
-
-import static j2html.TagCreator.document;
-import static j2html.TagCreator.html;
-import static org.assertj.core.api.Assertions.assertThat;
-import static play.test.Helpers.stubMessagesApi;
-
-import com.google.common.collect.ImmutableSet;
-import j2html.tags.specialized.DivTag;
-import junitparams.JUnitParamsRunner;
-import junitparams.Parameters;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import play.i18n.Lang;
-import play.i18n.Messages;
-import services.question.exceptions.UnsupportedQuestionTypeException;
-import services.question.types.QuestionType;
-import views.questiontypes.ApplicantQuestionRendererParams.ErrorDisplayMode;
-
-@RunWith(JUnitParamsRunner.class)
-public class ApplicantQuestionRendererFactoryTest {
-
-  private final Messages messages =
-      stubMessagesApi().preferred(ImmutableSet.of(Lang.defaultLang()));
-  private final ApplicantQuestionRendererParams params =
-      ApplicantQuestionRendererParams.builder()
-          .setMessages(messages)
-          .setErrorDisplayMode(ErrorDisplayMode.HIDE_ERRORS)
-          .build();
-
-  @Test
-  @Parameters(source = QuestionType.class)
-  public void rendererExistsForAllTypes(QuestionType type) throws UnsupportedQuestionTypeException {
-    // A null question type is not allowed to be created and won't show in the list
-    if (type == QuestionType.NULL_QUESTION
-        || type == QuestionType.MAP
-        || type == QuestionType.FILEUPLOAD) {
-      return;
-    }
-
-    ApplicantQuestionRendererFactory factory = new ApplicantQuestionRendererFactory();
-
-    ApplicantQuestionRenderer sampleRenderer = factory.getSampleRenderer(type);
-
-    DivTag content = sampleRenderer.render(params);
-    String renderedContent = document(html(content));
-    assertThat(renderedContent).contains("Sample question text");
-    assertThat(renderedContent).doesNotContain("help text");
-  }
-
-  @Test
-  @Parameters(source = QuestionType.class)
-  public void compositeQuestionsUseFieldset(QuestionType type)
-      throws UnsupportedQuestionTypeException {
-    // A null question type is not allowed to be created and won't show in the list
-    // Map type questions are only compatible with North Star so won't get rendered for this test
-    if (type == QuestionType.NULL_QUESTION
-        || type == QuestionType.MAP
-        || type == QuestionType.FILEUPLOAD) {
-      return;
-    }
-
-    // Multi-input questions should be wrapped in fieldsets for screen reader users.
-    ApplicantQuestionRendererFactory factory = new ApplicantQuestionRendererFactory();
-
-    ApplicantQuestionRenderer sampleRenderer = factory.getSampleRenderer(type);
-
-    DivTag content = sampleRenderer.render(params);
-    String renderedContent = document(html(content));
-    switch (type) {
-      case ADDRESS, CHECKBOX, ENUMERATOR, NAME, RADIO_BUTTON, YES_NO ->
-          assertThat(renderedContent).contains("fieldset");
-      case CURRENCY, DATE, DROPDOWN, EMAIL, ID, NUMBER, PHONE, STATIC, TEXT ->
-          assertThat(renderedContent).doesNotContain("fieldset");
-
-        // This is here because errorprone doesn't like that it was missing
-      case MAP, NULL_QUESTION, FILEUPLOAD -> {}
-    }
-  }
-}
