--- server/test/views/questiontypes/QuestionRendererTest.java
@@ -1,3 +1,0 @@
-package views.questiontypes;
-
-public abstract class QuestionRendererTest {}
